Update plugin to 2023.2.656-Dev
- Scope editor headers to WITH_EDITOR
- CavrnusPlugin/Plugins/CavrnusConnector/Source/CavrnusConnector/Private/CavrnusSpatialConnectorSubSystem.cpp

+++ Plugins/CavrnusConnector/Source/CavrnusConnector/Private/Translation/CavrnusProtoTranslation.cpp
@@ -23,7 +23,7 @@ namespace Cavrnus
 		return msg;
 	}
 
-	const ServerData::RelayClientMessage CavrnusProtoTranslation::BuildUpdateTimeMsg(float currTime)
+	const ServerData::RelayClientMessage CavrnusProtoTranslation::BuildUpdateTimeMsg(double currTime)
 	{
 		ServerData::UpdateTime t;
 		t.set_time(currTime);
